fix: call of undefined errors in delegate modules
