+++ package.json
@@ -33,7 +33,7 @@
     "prod": "nx build:libs && nx run-many --target=build --verbose --output-style=stream",
     "commit": "cz",
     "git:pruneTags": "git tag -l | xargs git tag -d && git fetch --all --prune --tags",
-    "prerelease:all": "nx run-many --target=version --projects=utils,node,nextjs-mf --releaseAs=prerelease --preid=rc --maxParallel=1 --allowEmptyRelease=true",
+    "prerelease:all": "nx run-many --target=version --projects=utils,node,nextjs-mf --maxParallel=1 --allowEmptyRelease=true",
     "release:node": "nx run node:version",
     "release:next": "nx run nextjs-mf:version",
     "release:utils": "nx run utils:version",
