+++ packages/nextjs-mf/src/default-delegate.ts
@@ -1,10 +1,7 @@
+import { importDelegatedModule } from '@module-federation/utilities/src/utils/importDelegatedModule'
+
 // eslint-disable-next-line no-async-promise-executor
 module.exports = new Promise(async (resolve, reject) => {
-  // @ts-ignore
-  const { importDelegatedModule } = await import(
-    // @ts-ignore
-    '@module-federation/utilities/src/utils/importDelegatedModule'
-  );
   // eslint-disable-next-line no-undef
   const currentRequest = new URLSearchParams(__resourceQuery).get('remote');
   // @ts-ignore
